Changed the update frequency to two seconds.  Raid Warning alert.
Realized that in a real loot roll situation 1 second updates is
way too fast and doesn't give people enough time to react.  Now
the timer updates every two seconds essentially doubling the
COUNT_DOWN_TIMER.

Also changed the announcement of a loot roll for an item to the
raid warning because some people may miss it if it doesn't flash
and make noise infront of them.

g0st