Rewrote CommandHandler() function and updated OnUpdate() function.
I added a check to see if the player is a raid leader or officer to
check whether or not the loot roll announcement should be displayed
in raid chat or as a raid warning.

I added a new variable countdown because COUNT_DOWN_TIMER was being
used as the counter and was not working as a constant configuration
variable.

Completely rewrote the CommandHandler() function and now it works
with "help" or no arguments at all.